Replace applesccript module, close #3
The external apple script module is generatinf and open a temp file per
applescript execution and does not close it. So we run in the issue #3

This change replace the applescript module by a very simple function
which does not suffer from this problem.

Also, some code and test cleanup